refactor(kd-tree): :recycle: migrate to the `go-kd-tree` package for the kd-tree data structure
This commit is responsible for migrating from copy-pasting code from the `go-kd-tree` (link: https://github.com/rishitc/go-kd-tree) project to adding it as a dependency to the `willow-go` project.
With this change in approach, new features, fixes, and performance improvements added to `go-kd-tree` will be readily available to the `willow-go` project.

Additionally, I have cleaned up all the associated codes and refactored them for this change. As part of this, the naming scheme for defining the `kdnode`'s `Key` has also been simplified, making it easier to understand and work with.

I have retained the `Query` function's interface in `willow-go`; however, I have entirely changed its implementation to depend on the `go-kd-tree` project's APIs. In the future, we can also clean up the `Query` function's interface and directly depend on the `go-kd-tree` project's APIs.
For now, I retain the `Query` function's interface to avoid excessive code breakages and refactors.

Reference:
1. `go-kd-tree` Project: https://github.com/rishitc/go-kd-tree